requirements.txt
python-dotenv
langchain_openai